refactor(All): 전부다 수정

+++ src/main/java/lotto/dto/WinningLottoDto.java
@@ -2,5 +2,5 @@ package lotto.dto;
 
 import java.util.List;
 
-public record WinningLottoDto(List<Integer> winningNumbers, long bonusNumber) {
+public record WinningLottoDto(List<Integer> winningNumbers, int bonusNumber) {
 }
